Add UsersController and RoleEnum; refactor AuthController to use RoleEnum for role checks
