+++ src/main/java/edu/harvard/iq/dataverse/marketplace/model/enums/RoleEnum.java
@@ -1,0 +1,18 @@
+package edu.harvard.iq.dataverse.marketplace.model.enums;
+
+public enum RoleEnum {
+
+    ADMIN("ADMIN"),
+    USER("USER");
+
+    private final String name;
+
+    RoleEnum(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+}
